Add trait MassaJsonRpc (#18)
* Add trait MassaJsonRpc
* Implement trait MassaJsonRpc with jsonrpsee & reqwest
