+++ hello_world_scripts/examples/get_status.rs
@@ -1,19 +1,11 @@
-use massa_rust_web3::{
-    BUILDNET_URL,
-    get_status,
-    // get_status_reqwest
-};
+use massa_rust_web3::{BUILDNET_URL, MassaRpcClient, MassaJsonRpc};
 
 #[tokio::main]
 async fn main() {
-    let node_status = get_status(BUILDNET_URL).await.unwrap();
+
+    let client = MassaRpcClient::new(BUILDNET_URL);
+    let node_status = client.get_status().await.unwrap();
     println!("{}", "#".repeat(20));
     println!("Node status: {}", node_status);
     println!("{}", "#".repeat(35));
-
-    /*
-    let node_status = get_status_reqwest(BUILDNET_URL).await.unwrap();
-    println!("{}", "#".repeat(20));
-    println!("Node status: {}", node_status);
-    */
 }
