+++ massa_rust_web3/Cargo.toml
@@ -4,12 +4,22 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-jsonrpsee = { version = "0.24", features = ["full"] }
-serde = { version = "1.0.219", features = ["derive"] }
+serde = { version = "1", features = ["derive"] }
+thiserror = "2"
+tracing = "0.1"
+tokio = { version = "1" }
+bitte = { version = "0.0.1", features = ["threads"] }
+# massa deps
 massa_api_exports = { git = "https://github.com/massalabs/massa", package = "massa_api_exports" }
 massa_models = { git = "https://github.com/massalabs/massa", package = "massa_models" }
 massa_signature = { git = "https://github.com/massalabs/massa", package = "massa_signature" }
 massa_serialization = { git = "https://github.com/massalabs/massa", package = "massa_serialization" }
-reqwest = { version = "0.12.23", features = ["json"] }
-thiserror = "2.0"
-tracing = "0.1"
+# optional deps
+reqwest = { version = "0.12", default-features = false, features = ["default-tls", "json"], optional = true }
+serde_json = { version = "1", optional = true }
+jsonrpsee = { version = "0.26", features = ["full"], optional = true }
+
+[features]
+default = ["jsonrpsee"]
+jsonrpsee = ["dep:jsonrpsee"]
+reqwest = ["dep:reqwest", "dep:serde_json"]
